requirements.txt
setuptools
pandas
numpy
scikit-learn
seaborn
python-dotenv
mysql-connector-python
pymysql
catboost
xgboost
streamlit

# -e .